new_backend/requirements.txt
fastapi
uvicorn[standard]
supabase
pydantic
python-dotenv
google-generativeai